rename Nav to NavBar
