+++ src/components/index.tsx
@@ -1,8 +1,8 @@
-import Nav from "./Nav/Nav";
-import { EmailSignupForm } from "./EmailSignupForm/EmailSignupForm";
-import { FeedCard } from "./FeedCard/FeedCard";
-import { ListRow } from "./Row/Row";
-import { Link } from "./Link/Link";
-import { Icon } from "./Icon/Icon";
+import Nav from './NavBar/NavBar'
+import {EmailSignupForm} from './EmailSignupForm/EmailSignupForm'
+import {FeedCard} from './FeedCard/FeedCard'
+import {ListRow} from './Row/Row'
+import {Link} from './Link/Link'
+import {Icon} from './Icon/Icon'
 
-export { Nav, EmailSignupForm, FeedCard, ListRow, Icon, Link };
+export {Nav, EmailSignupForm, FeedCard, ListRow, Icon, Link}
